+++ packages/taro-transformer-wx/__tests__/template.spec.ts
@@ -214,7 +214,10 @@ describe('Template', () => {
         // const props = instance.$props.Custom()
         // expect(props.$name).toBe('Custom')
         // expect(props.hidden).toBe(true)
-        expect(template).toMatch(`<template is=\"Custom\" data=\"{{...$$Custom}}\" wx:for-item=\"item\"></template>`)
+        expect(template).toMatch(`data=\"{{...item}}\"`)
+        expect(template).toMatch(`wx:for-item=\"item\"`)
+        expect(template).toMatch(`wx:for=\"{{$$Custom}}\"`)
+        expect(template).toMatch(`wx:key=`)
       })
 
       test('自定义组件循环', () => {
@@ -249,75 +252,6 @@ describe('Template', () => {
       })
 
       expect(template).toMatch('<view hover-class="test"></view>')
-    })
-
-    describe('if statement', () => {
-      test('简单情况', () => {
-        const { template, ast } = transform({
-          ...baseOptions,
-          isRoot: true,
-          code: buildComponent(`
-          const tasks = []
-          if (tasks !== null) {
-            return <View className='page-body' >
-            </View>
-          }
-
-          return (
-            <View className='page-body'>
-              <Text>Hello world!</Text>
-            </View>
-          )
-          `)
-        })
-
-        expect(template).toMatch(prettyPrint(`
-        <block>
-            <view class=\"page-body\" wx:if=\"{{tasks !== null}}\"></view>
-            <view class=\"page-body\" wx:else>
-                <text>Hello world!</text>
-            </view>
-        </block>
-        `))
-      })
-
-      test('两个平级的 ifStatement', () => {
-        const { template, ast } = transform({
-          ...baseOptions,
-          isRoot: true,
-          code: buildComponent(`
-          const tasks = []
-          if (tasks !== null) {
-            return <View className='page-body' >
-            </View>
-          }
-
-          if (tasks.length === 0) {
-            return <View className='page-body'>
-              <Text>{tasks.length}</Text>
-            </View>
-          }
-
-          return (
-            <View className='page-body'>
-              <Text>Hello world!</Text>
-            </View>
-          )
-          `)
-        })
-
-        expect(template).toMatch(prettyPrint(`
-          <block>
-              <view class=\"page-body\" wx:if=\"{{tasks !== null}}\"></view>
-              <view class=\"page-body\" wx:elif=\"{{tasks.length === 0}}\">
-                  <text>{{tasks.length}}</text>
-              </view>
-              <view class=\"page-body\" wx:else>
-                  <text>Hello world!</text>
-              </view>
-          </block>
-        `))
-      })
     })
 
     describe('JSX 元素引用', () => {